requirements.txt
# Google API Client Libraries
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Google Generative AI (Gemini)
google-generativeai>=0.4.0

# Optional: If you prefer OpenAI instead of Gemini
# openai==1.6.1

# Core dependencies
httplib2==0.22.0
uritemplate==4.1.1
protobuf==4.25.1

# Email handling
email-validator==2.1.0

# Environment variable management
python-dotenv==1.0.0

# Timezone handling
pytz==2023.3

# Contract Agent Dependencies
# Vector Database
chromadb>=0.4.15

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.0
python-docx==0.8.11
pytesseract==0.3.10  # OCR fallback

# Text Processing & Similarity
rapidfuzz==3.5.2
nltk==3.8.1

# Report Generation
jinja2==3.1.2
weasyprint>=60.0
reportlab==4.0.7

# Data Processing
pandas>=2.1.3,<3.0.0
numpy>=1.26.0,<2.0.0

# Utilities
tqdm==4.66.1
typing-extensions==4.8.0

# Web Framework Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1 